+++ packages/f-searchbox/src/components/tests/Form.test.js
@@ -113,6 +113,34 @@ describe('`Form`', () => {
                 });
 
                 describe('when `hasLastSavedAddress` is `truthy`', () => {
+                    it('should `preventDefault` so the we can post a payload correctly', () => {
+                        // Arrange
+                        const address = 'AR511AR';
+
+                        const propsData = {
+                            config: {
+                                address,
+                                locationFormat: () => jest.fn()
+                            },
+                            service: {
+                                isValid: jest.fn(() => true)
+                            }
+                        };
+                        const wrapper = shallowMount(Form, {
+                            propsData,
+                            store: createStore(),
+                            localVue
+                        });
+
+                        wrapper.setData({ address, lastAddress: address });
+
+                        // Act
+                        wrapper.vm.submit(event);
+
+                        // Assert
+                        expect(event.preventDefault).toHaveBeenCalled();
+                    });
+
                     it('should make a call to `searchPreviouslySavedAddress` with `event`', () => {
                         // Arrange
                         const address = 'AR511AR';
@@ -436,7 +464,6 @@ describe('`Form`', () => {
                         onSubmit: false,
                         formUrl: 'search/Andromeda',
                         form: '<form></form>',
-                        callback: getLastLocation,
                         event
                     };
 
@@ -444,7 +471,7 @@ describe('`Form`', () => {
                     wrapper.vm.searchPreviouslySavedAddress(event);
 
                     // Assert
-                    expect(spy).toHaveBeenCalledWith(searchPayload);
+                    expect(spy).toHaveBeenCalledWith(searchPayload, getLastLocation());
                 });
             });
         });